Further small refactoring/cleanup
