+++ tests/system/include/build_cxx/system_tests/env.hxx
@@ -23,8 +23,9 @@
 
 #include "build_cxx/system_tests/test_env.hxx"
 
-namespace build_cxx::system_tests {
+namespace build_cxx::system_tests::impl {
 
+// Not to be used in test cases, it's only purpose is initialization of test_env
 struct env {
   [[nodiscard]] static env &inst() noexcept;
