+++ tests/system/src/test_env.cxx
@@ -25,7 +25,7 @@ namespace build_cxx::system_tests {
 
 test_env const &test_env::inst() noexcept {
   // force 2 lines
-  return env::inst().te;
+  return impl::env::inst().te;
 }
 
 test_env::test_env() noexcept = default;
